# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "La Mesa API",
    "description": "REST JSON contract for **La Mesa**. Every mutating response carries a full A2UI\nmessage array. No WebSockets or SSE (INV-010).\n\nFrozen routes live in `SPECS.md` §8 (`REQ-API-01` … `REQ-API-08`). The frontend\nconsumes this HTTP API only; LLM providers are not called from the client.\n\nInteractive docs: `/swagger` or `/docs` (Swagger UI) and `/redoc`. Machine-readable spec:\n`/openapi.json`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/action": {
      "post": {
        "tags": [
          "action"
        ],
        "summary": "Submit closed-loop A2UI action (REQ-API-03)",
        "description": "Submit a user interaction emitted by an A2UI component (REQ-API-03, INV-017).\n\n- Closes the loop: interactions such as moving a liquidity/term slider (`tune_tradeoff`),\n  toggling assumptions (`toggle_assumption`), or accepting a debt consolidation offer (`accept_offer`)\n  are fed back into the agent context.\n- If accepting an offer, launches or transitions to the El Revés negotiation loop.\n- If tuning tradeoffs or assumptions, recalculates the deterministic financial math via MCP\n  and triggers structural revalidation or BreakAlert mutation.",
        "operationId": "post_action",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Action processed successfully; returns updated A2UI message array.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Surface not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/negotiation/{session_id}/turn": {
      "post": {
        "tags": [
          "negotiation"
        ],
        "summary": "Advance El Revés negotiation turn (REQ-API-06)",
        "description": "Execute one round of the El Revés bank negotiation (REQ-API-06, INV-001).\n\n- The agent negotiates against a simulated bank persona using bank lender policies.\n- Evaluates interest rates, terms, and liquidity tradeoff constraints.\n- Emits updated negotiation UI with bank counteroffer or final agreement.",
        "operationId": "negotiation_turn",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Round evaluated; returns updated offer state and negotiation A2UI.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/negotiation/{session_id}/take-control": {
      "post": {
        "tags": [
          "negotiation"
        ],
        "summary": "User takes manual control of El Revés negotiation (REQ-API-06)",
        "description": "Allow the user to intervene directly and propose custom terms to the bank persona.",
        "operationId": "negotiation_take_control",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manual position evaluated by the bank persona.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "surface_id": {
            "type": "string",
            "title": "Surface Id",
            "description": "ID of the generated UI surface where the action originated.",
            "examples": [
              "surf_abc123"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "A2UI action name, e.g. tune_tradeoff, toggle_assumption, accept_offer (INV-017).",
            "examples": [
              "tune_tradeoff"
            ]
          },
          "source_component_id": {
            "type": "string",
            "title": "Source Component Id",
            "description": "ID of the A2UI component that dispatched the action.",
            "default": "",
            "examples": [
              "tradeoff_slider"
            ]
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context",
            "description": "Arbitrary action context parameters provided by the component.",
            "examples": [
              {
                "liquidity_weight": 0.8
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "surface_id",
          "name"
        ],
        "title": "ActionRequest",
        "description": "Closed-loop action payload emitted from an A2UI component.",
        "examples": [
          {
            "context": {
              "liquidity_weight": 0.8
            },
            "name": "tune_tradeoff",
            "source_component_id": "tradeoff_slider",
            "surface_id": "surf_abc123"
          }
        ]
      },
      "AgentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status ('ok' or 'error').",
            "examples": [
              "ok"
            ]
          },
          "surface_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface Id",
            "description": "Identifier of the active or newly persisted A2UI surface.",
            "examples": [
              "surf_1a2b3c4d"
            ]
          },
          "a2ui": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "A2Ui",
            "description": "Full A2UI message array for this turn (INV-010, INV-016)."
          },
          "assistant_text": {
            "type": "string",
            "title": "Assistant Text",
            "description": "Text utterance or commentary from the AI agent.",
            "default": "",
            "examples": [
              "He analizado tus deudas y reestructurado el plan."
            ]
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues",
            "description": "List of warnings, schema validation issues, or soft errors."
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Optional human-readable error or status message."
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor",
            "description": "Current acting party in negotiation ('user' or 'bank').",
            "examples": [
              "bank"
            ]
          },
          "round": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Round",
            "description": "Current negotiation round index.",
            "examples": [
              1
            ]
          },
          "catalog_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Id",
            "description": "Catalog ID used to render the interface ('amitie.standard.v1' or 'amitie.voz_color.v1').",
            "examples": [
              "amitie.standard.v1"
            ]
          },
          "audio_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Ref",
            "description": "URL to stream generated TTS audio for voice accessibility (REQ-ACC-03).",
            "examples": [
              "/api/audio/asset_123"
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "Structured error code (e.g. 'provider_unavailable', 'bad_request')."
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "description": "Indicates whether the client may retry this turn (REQ-NFR-04).",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AgentResponse",
        "description": "Standard agent turn response carrying the generated A2UI interface and metadata.",
        "examples": [
          {
            "a2ui": [
              {
                "catalogId": "amitie.standard.v1",
                "root": "root",
                "surfaceId": "surf_1a2b3c4d",
                "type": "beginRendering"
              }
            ],
            "assistant_text": "He analizado tus deudas y reestructurado el plan.",
            "audio_ref": "/api/audio/asset_123",
            "catalog_id": "amitie.standard.v1",
            "issues": [],
            "retryable": false,
            "status": "ok",
            "surface_id": "surf_1a2b3c4d"
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HttpErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation of why the request failed.",
            "examples": [
              "unknown session"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HttpErrorDetail",
        "description": "Standard HTTP error response payload.",
        "examples": [
          {
            "detail": "unknown session"
          }
        ]
      },
...
      "NegotiationRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier negotiating terms.",
            "default": "u_ana",
            "examples": [
              "u_ana"
            ]
          },
          "position": {
            "additionalProperties": true,
            "type": "object",
            "title": "Position",
            "description": "User-proposed terms or counteroffer parameters."
          }
        },
        "type": "object",
        "title": "NegotiationRequest",
        "description": "Payload to advance or intervene in the El Revés negotiation loop.",
        "examples": [
          {
            "position": {
              "proposed_payment": 2800.0,
              "proposed_rate": 0.22,
              "proposed_term_months": 36
            },
            "user_id": "u_ana"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}